Move emission_source.py general methods to oil and gas sector root
The emission_sources folder has become a home for integration methods which extract emission sources from various regional datasets. This file contains the general template and helper methods, so makes sense be closer to the module root.
