+++ src/openmethane_prior/sectors/oil_gas/emission_source.py
